requirements.txt
django-toolbelt==0.0.1
dj-database-url==0.3.0
django-ajaximage==0.2.7
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
django-endless-pagination==2.0.0
